requirements.txt
Flask>=2.0
Flask-Login>=0.6
Flask-MySQLdb>=1.0.1
python-dotenv>=1.0
werkzeug>=2.0
